Cabal modules & states support user installs
For security reasons we don't want to have to run all cabal software as
root so we need to be able to run commands as root.  This updates the
cabal modules and states to use the the better internal salt command
API which lets us pick the user to run the command as.